Add preferences concept with example usage
Figured this was a bit easier than having to write this all yourself
every time plus it promotes preferences to the application level.

The only other thing I might think of doing would be to save
preferences by default when an application is closing to mirror the
automatic loading of saved preferences.