Make grid monitor use scaled drawImage for loupe
The grid monitor now uses the working canvas as scaled up without
image smoothing to "zoom" in on the image. When we can't disable the
image smoothing the old approach is used and the zoom level is locked
at 10.

This appears to be less intensive, and is certainly more flexible.

Also we again defer colorPick by some time to reduce canvas calls.