Reduce updates to download image
While it typically isn't a problem, there's no need to get the canvas
information on every mouseover of the download thumbnail.

Now we keep a flag for whether the download image is stale and
only update on mouseovers when that's the case.

Ideally we'll switch over to File API, though this method is more
backward compatible.