Improve color picking with workers and promises
When possible we transfer the canvas data to a worker and get pixel data
from the worker when possible.

There's probably a lot more throttling we could add to this but for
a first pass it's at least working. The PhotoEditor's dataAtPoint
API has been updated to always return a promise regardless of whether
we're using a worker for consistency.

It may be worth creating multiple workers as needed as well as
relegating more canvas operations to workers.