Pre-render PhotoFX loupe grid on foreground canvas
We only need to draw it every so often, might as well let the GPU
composite it. Though we do miss out on some of the fun drawing modes,
it should end up being a bit faster.

This also adds the very experimental largeCanvas zooming which
should ultimately be the fastest way to zoom in as the zoom itself is
pre-rendered and simply slid into place. Of course the resultant
canvas is huge and needs to be tiled to do this nicely.

Also tried to improve functionally of gridMagnifier in Firefox.

The largeCanvas loupe masking requires a real mask as the border-radius
only worked when the canvas was the expected size; I added an SVG
circle mask.